Fix several bugs and delete unnecessary files
