--- requirements.txt
@@ -1,2 +1,0 @@
-beautifulsoup4==4.3.2
-requests==2.5.1
